feat: move download into bank transactions menu; add record-expense (+) button
Move the standalone Download button into the header hamburger menu in all
views, and add a '+' record-expense icon button to its left that opens a
stub modal. Extract SelectedBankAccountsChip (with its CSS) and
TransactionsSearch into their own component files.
