+++ src/components/BankTransactions/SelectedBankAccountsChip/selectedBankAccountsChip.scss
@@ -1,0 +1,13 @@
+.Layer__bank-transactions__selected-accounts-chip--wide {
+  display: none;
+
+  @container (min-width: 500px) {
+    display: block;
+  }
+}
+
+.Layer__bank-transactions__selected-accounts-chip--compact {
+  @container (min-width: 500px) {
+    display: none;
+  }
+}
