+++ src/components/BankTransactions/bankTransactionsHeader.scss
@@ -64,17 +64,3 @@
     box-shadow: none;
   }
 }
-
-.Layer__bank-transactions__selected-accounts-chip--wide {
-  display: none;
-
-  @container (min-width: 500px) {
-    display: block;
-  }
-}
-
-.Layer__bank-transactions__selected-accounts-chip--compact {
-  @container (min-width: 500px) {
-    display: none;
-  }
-}
